* all about integer conversions [rtj]

+++ xstream/src/bz.cpp
@@ -128,7 +128,7 @@ namespace bz {
         z_strm->bzfree = NULL;
         z_strm->opaque = NULL;
         //buffers
-        z_strm->avail_out = out.size;
+        z_strm->avail_out = (unsigned int)out.size;
         z_strm->next_out = out.buf;
 
         z_strm->avail_in = 0;
@@ -270,7 +270,7 @@ namespace bz {
     std::streamsize ostreambuf::xsputn(const char *buffer, std::streamsize n) {
         LOG("bz::ostreambuf::xsputn(" << buffer << "," << n << ")");
 
-        return flush(no_sync, buffer, n);
+        return flush(no_sync, buffer, (int)n);
     }
 
     int ostreambuf::flush(flush_kind f, const char *appendbuf, int appendsize) {
@@ -282,8 +282,8 @@ namespace bz {
         int written;
         if (in_s > 0) {
             z_strm->next_in = pbase();
-            z_strm->avail_in = in_s;
-            written = in_s;
+            z_strm->avail_in = (unsigned int)in_s;
+            written = (int)in_s;
         } else if (appendsize > 0) {
             z_strm->next_in = (char*)appendbuf;
             z_strm->avail_in = appendsize;
@@ -364,7 +364,7 @@ namespace bz {
                 std::streamsize count = out.size - z_strm->avail_out;
                 if (count > 0) {  // ignore empty blocks
                     LOG("\twriting " << count << " bytes");
-                    int size = htonl(count);
+                    int size = htonl((unsigned int)count);
                     MUTEX_LOCK
                     const std::streamsize wrote = _sb->sputn((char*)&size, 4) +
                                                   _sb->sputn(out.buf, count);
@@ -380,7 +380,7 @@ namespace bz {
                     MUTEX_UNLOCK
                 }
                 z_strm->next_out = out.buf;
-                z_strm->avail_out = out.size;
+                z_strm->avail_out = (unsigned int)out.size;
             }
 
             if ((0 == z_strm->avail_out) && (0 != z_strm->avail_in)) {
@@ -408,7 +408,7 @@ namespace bz {
             z_strm->bzalloc = NULL;
             z_strm->bzfree = NULL;
             z_strm->opaque = NULL;
-            z_strm->avail_out = out.size;
+            z_strm->avail_out = (unsigned int)out.size;
             z_strm->next_out = out.buf;
             z_strm->avail_in = 0;
             z_strm->next_in = in.buf;
@@ -490,9 +490,9 @@ namespace bz {
             }
             while (block_offset < new_block_offset) {
                 z_strm->next_out = out.buf;
-                z_strm->avail_out = new_block_offset - block_offset;
+                z_strm->avail_out = (unsigned int)(new_block_offset - block_offset);
                 if (z_strm->avail_out > out.size) {
-                    z_strm->avail_out = out.size;
+                    z_strm->avail_out = (unsigned int)out.size;
                 }
                 block_offset += z_strm->avail_out;
                 decompress();
@@ -501,7 +501,7 @@ namespace bz {
             new_block_offset = 0;
         }
 
-        z_strm->avail_out = out.size;
+        z_strm->avail_out = (unsigned int)out.size;
         z_strm->next_out = out.buf;
         if (0 < z_strm->avail_in) {
             LOG("\tdata in queue, inflating");
@@ -539,8 +539,8 @@ namespace bz {
             else
             {
                 std::streamsize available = egptr() - gptr();
-                int waste = new_block_offset - block_offset;
-                waste = (available < waste)? available : waste;
+                int waste = (int)(new_block_offset - block_offset);
+                waste = (available < waste)? (int)available : waste;
                 if (waste > 0) {
                     gbump(waste);
                     block_offset += waste;
@@ -548,9 +548,9 @@ namespace bz {
             }
             while (block_offset < new_block_offset) {
                 z_strm->next_out = out.buf;
-                z_strm->avail_out = new_block_offset - block_offset;
+                z_strm->avail_out = (unsigned int)(new_block_offset - block_offset);
                 if (z_strm->avail_out > out.size) {
-                    z_strm->avail_out = out.size;
+                    z_strm->avail_out = (unsigned int)out.size;
                 }
                 block_offset += z_strm->avail_out;
                 decompress();
@@ -561,7 +561,7 @@ namespace bz {
 
         //try to satisfy request from buffered input
         std::streamsize available = egptr() - gptr();
-        int read = (available >= n)? n : available;
+        int read = (available >= n)? n : (int)available;
         if (read) {
             std::copy(gptr(), gptr() + read, buffer);
             gbump(read);
@@ -577,7 +577,7 @@ namespace bz {
             }
 
             z_strm->next_out = buffer + read;
-            z_strm->avail_out = n - read;
+            z_strm->avail_out = (unsigned int)(n - read);
 
             if (0 < z_strm->avail_in) {
                 decompress();
@@ -598,7 +598,7 @@ namespace bz {
     void istreambuf::read_decompress() {
         LOG("bz::istreambuf::read_decompress ");
         bool reinit_decompressor = false;
-        int read;
+        size_t read;
         if (block_size < 0) { // stream has no blocksize markers
             MUTEX_LOCK
             if (new_block_start > 0) {
